api/requirements.txt
Flask==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.4.4
Flask-Cors==4.0.0
Flasgger==0.9.7
PyMySQL==1.1.0
python-dotenv==1.0.0
Werkzeug==3.0.1
bcrypt==4.1.2
